chore: refactor layout
